# Relyf.Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Relyf API",
    "description": "API for the Relyf upcycling platform - connecting users with creative ways to reuse and repurpose items",
    "version": "v1",
    "contact": {
      "name": "Relyf Team",
      "url": "https://github.com/nanasarf/Relyf"
    }
  },
  "servers": [
    {
      "url": "https://localhost:7139",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/Projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List user's projects",
        "description": "Get paginated list of projects for authenticated user",
        "operationId": "ListProjects",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedProjectsDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a new project",
        "description": "Create a new upcycling project (requires authentication)",
        "operationId": "CreateProject",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/Projects/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get project by ID",
        "description": "Retrieve project details with steps (requires authentication and ownership)",
        "operationId": "GetProject",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectWithStepsDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
    "/api/Projects/{id}/steps": {
      "put": {
        "tags": ["Projects"],
        "summary": "Update project steps",
        "description": "Replace all steps for a project (requires authentication and ownership)",
        "operationId": "UpdateProjectSteps",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertStepsRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Steps updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
    "/api/Projects/{id}/status": {
      "patch": {
        "tags": ["Projects"],
        "summary": "Update project status",
        "description": "Change project status (draft, in_progress, completed)",
        "operationId": "UpdateProjectStatus",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Status updated successfully"
          },
          "400": {
            "description": "Invalid status value"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "type": "object",
        "required": ["userId", "title"],
        "properties": {
          "userId": {
            "type": "integer",
            "description": "User ID (will be overridden by authenticated user)"
          },
          "ideaId": {
            "type": "integer",
            "nullable": true,
            "description": "Optional: Link to a user-created idea"
          },
          "aiIdeaId": {
            "type": "integer",
            "nullable": true,
            "description": "Optional: Link to an AI-generated idea"
          },
          "title": {
            "type": "string",
            "example": "Plastic Bottle Planter"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Turn plastic bottles into hanging planters"
          }
        }
      },
      "ProjectDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "integer"
          },
          "ideaId": {
            "type": "integer",
            "nullable": true
          },
          "aiIdeaId": {
            "type": "integer",
            "nullable": true
          },
          "userId": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["draft", "in_progress", "completed"],
            "description": "Project status"
          }
        }
      },
      "ProjectWithStepsDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProjectDto"
          },
          {
            "type": "object",
            "properties": {
              "steps": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StepDto"
                }
              }
            }
          }
        ]
      },
      "StepDto": {
        "type": "object",
        "properties": {
          "projectStepId": {
            "type": "integer"
          },
          "stepNumber": {
            "type": "integer"
          },
          "instruction": {
            "type": "string"
          }
        }
      },
      "PagedProjectsDto": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectDto"
            }
          },
          "total": {
            "type": "integer"
          },
          "skip": {
            "type": "integer"
          },
          "take": {
            "type": "integer"
          }
        }
      },
      "UpsertStepsRequest": {
        "type": "object",
        "required": ["steps"],
        "properties": {
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of step instructions in order",
            "example": [
              "Cut the bottle in half",
              "Punch drainage holes in the bottom",
              "Fill with soil and plant seeds"
            ]
          }
        }
      },
      "UpdateStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["draft", "in_progress", "completed"],
            "example": "in_progress"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      }
    }
  },
...
}